fixes and move metadata and validations out
* Move Metadata and Validation function to separate files
* fix a few minor summary code issues
* Remove files from the summary stats - not much point because it is accessible from the mbtiles struct itself
