+++ mbtiles/src/mbtiles.rs
@@ -1,60 +1,22 @@
 #![allow(clippy::missing_errors_doc)]
 
-use std::collections::HashSet;
 use std::ffi::OsStr;
 use std::fmt::{Display, Formatter};
 use std::path::Path;
-use std::str::FromStr;
 
 #[cfg(feature = "cli")]
 use clap::ValueEnum;
 use enum_display::EnumDisplay;
-use futures::TryStreamExt;
 use log::{debug, info, warn};
 use martin_tile_utils::{Format, TileInfo};
-use serde::ser::SerializeStruct;
-use serde::{Serialize, Serializer};
-use serde_json::{Value as JSONValue, Value};
+use serde_json::Value;
 use sqlite_hashes::register_md5_function;
-use sqlx::sqlite::{SqliteConnectOptions, SqliteRow};
-use sqlx::{query, Connection as _, Row, SqliteConnection, SqliteExecutor};
-use tilejson::{tilejson, Bounds, Center, TileJSON};
+use sqlx::sqlite::SqliteConnectOptions;
+use sqlx::{query, Connection as _, SqliteConnection, SqliteExecutor};
+use tilejson::TileJSON;
 
 use crate::errors::{MbtError, MbtResult};
-use crate::queries::{
-    has_tiles_with_hash, is_flat_tables_type, is_flat_with_hash_tables_type,
-    is_normalized_tables_type,
-};
-use crate::MbtError::{
-    AggHashMismatch, AggHashValueNotFound, FailedIntegrityCheck, IncorrectTileHash,
-};
-
-#[derive(Clone, Debug, PartialEq, Serialize)]
-pub struct Metadata {
-    pub id: String,
-    #[serde(serialize_with = "serialize_ti")]
-    pub tile_info: TileInfo,
-    pub layer_type: Option<String>,
-    pub tilejson: TileJSON,
-    pub json: Option<JSONValue>,
-}
-
-fn serialize_ti<S: Serializer>(ti: &TileInfo, serializer: S) -> Result<S::Ok, S::Error> {
-    let mut s = serializer.serialize_struct("TileInfo", 2)?;
-    s.serialize_field("format", &ti.format.to_string())?;
-    s.serialize_field(
-        "encoding",
-        ti.encoding.content_encoding().unwrap_or_default(),
-    )?;
-    s.end()
-}
-
-/// Metadata key for the aggregate tiles hash value
-pub const AGG_TILES_HASH: &str = "agg_tiles_hash";
-
-/// Metadata key for a diff file,
-/// describing the eventual [`AGG_TILES_HASH`] value once the diff is applied
-pub const AGG_TILES_HASH_IN_DIFF: &str = "agg_tiles_hash_after_apply";
+use crate::MbtType;
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, EnumDisplay)]
 #[enum_display(case = "Kebab")]
@@ -65,37 +27,9 @@ pub enum MbtTypeCli {
     Normalized,
 }
 
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, EnumDisplay, Serialize)]
-#[enum_display(case = "Kebab")]
-pub enum MbtType {
-    Flat,
-    FlatWithHash,
-    Normalized { hash_view: bool },
-}
-
-impl MbtType {
-    pub fn is_normalized(&self) -> bool {
-        matches!(self, Self::Normalized { .. })
-    }
-
-    pub fn is_normalized_with_view(&self) -> bool {
-        matches!(self, Self::Normalized { hash_view: true })
-    }
-}
-
-#[derive(PartialEq, Eq, Default, Debug, Clone, EnumDisplay)]
-#[enum_display(case = "Kebab")]
-#[cfg_attr(feature = "cli", derive(ValueEnum))]
-pub enum IntegrityCheckType {
-    #[default]
-    Quick,
-    Full,
-    Off,
-}
-
 #[derive(Clone, Debug)]
 pub struct Mbtiles {
-    pub(crate) filepath: String,
+    filepath: String,
     filename: String,
 }
 
@@ -159,17 +93,6 @@ impl Mbtiles {
         &self.filename
     }
 
-    fn to_val<V, E: Display>(&self, val: Result<V, E>, title: &str) -> Option<V> {
-        match val {
-            Ok(v) => Some(v),
-            Err(err) => {
-                let name = &self.filename;
-                warn!("Unable to parse metadata {title} value in {name}: {err}");
-                None
-            }
-        }
-    }
-
     /// Attach this `MBTiles` file to the given `SQLite` connection as a given name
     pub async fn attach_to<T>(&self, conn: &mut T, name: &str) -> MbtResult<()>
     where
@@ -183,146 +106,7 @@ impl Mbtiles {
         Ok(())
     }
 
-    /// Get a single metadata value from the metadata table
-    pub async fn get_metadata_value<T>(&self, conn: &mut T, key: &str) -> MbtResult<Option<String>>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let query = query!("SELECT value from metadata where name = ?", key);
-        let row = query.fetch_optional(conn).await?;
-        if let Some(row) = row {
-            if let Some(value) = row.value {
-                return Ok(Some(value));
-            }
-        }
-        Ok(None)
-    }
-
-    pub async fn validate(
-        &self,
-        check_type: IntegrityCheckType,
-        update_agg_tiles_hash: bool,
-    ) -> MbtResult<String> {
-        let mut conn = if update_agg_tiles_hash {
-            self.open().await?
-        } else {
-            self.open_readonly().await?
-        };
-        self.check_integrity(&mut conn, check_type).await?;
-        self.check_each_tile_hash(&mut conn).await?;
-        if update_agg_tiles_hash {
-            self.update_agg_tiles_hash(&mut conn).await
-        } else {
-            self.check_agg_tiles_hashes(&mut conn).await
-        }
-    }
-
-    /// Get the aggregate tiles hash value from the metadata table
-    pub async fn get_agg_tiles_hash<T>(&self, conn: &mut T) -> MbtResult<Option<String>>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        self.get_metadata_value(&mut *conn, AGG_TILES_HASH).await
-    }
-
-    pub async fn set_metadata_value<T>(
-        &self,
-        conn: &mut T,
-        key: &str,
-        value: Option<&str>,
-    ) -> MbtResult<()>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        if let Some(value) = value {
-            query!(
-                "INSERT OR REPLACE INTO metadata(name, value) VALUES(?, ?)",
-                key,
-                value
-            )
-            .execute(conn)
-            .await?;
-        } else {
-            query!("DELETE FROM metadata WHERE name=?", key)
-                .execute(conn)
-                .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn get_metadata<T>(&self, conn: &mut T) -> MbtResult<Metadata>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let (tj, layer_type, json) = self.parse_metadata(conn).await?;
-
-        Ok(Metadata {
-            id: self.filename.to_string(),
-            tile_info: self.detect_format(&tj, conn).await?,
-            tilejson: tj,
-            layer_type,
-            json,
-        })
-    }
-
-    async fn parse_metadata<T>(
-        &self,
-        conn: &mut T,
-    ) -> MbtResult<(TileJSON, Option<String>, Option<Value>)>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let query = query!("SELECT name, value FROM metadata WHERE value IS NOT ''");
-        let mut rows = query.fetch(conn);
-
-        let mut tj = tilejson! { tiles: vec![] };
-        let mut layer_type: Option<String> = None;
-        let mut json: Option<JSONValue> = None;
-
-        while let Some(row) = rows.try_next().await? {
-            if let (Some(name), Some(value)) = (row.name, row.value) {
-                match name.as_ref() {
-                    "name" => tj.name = Some(value),
-                    "version" => tj.version = Some(value),
-                    "bounds" => tj.bounds = self.to_val(Bounds::from_str(value.as_str()), &name),
-                    "center" => tj.center = self.to_val(Center::from_str(value.as_str()), &name),
-                    "minzoom" => tj.minzoom = self.to_val(value.parse(), &name),
-                    "maxzoom" => tj.maxzoom = self.to_val(value.parse(), &name),
-                    "description" => tj.description = Some(value),
-                    "attribution" => tj.attribution = Some(value),
-                    "type" => layer_type = Some(value),
-                    "legend" => tj.legend = Some(value),
-                    "template" => tj.template = Some(value),
-                    "json" => json = self.to_val(serde_json::from_str(&value), &name),
-                    "format" | "generator" => {
-                        tj.other.insert(name, Value::String(value));
-                    }
-                    _ => {
-                        let file = &self.filename;
-                        info!("{file} has an unrecognized metadata value {name}={value}");
-                        tj.other.insert(name, Value::String(value));
-                    }
-                }
-            }
-        }
-
-        if let Some(JSONValue::Object(obj)) = &mut json {
-            if let Some(value) = obj.remove("vector_layers") {
-                if let Ok(v) = serde_json::from_value(value) {
-                    tj.vector_layers = Some(v);
-                } else {
-                    warn!(
-                        "Unable to parse metadata vector_layers value in {}",
-                        self.filename
-                    );
-                }
-            }
-        }
-
-        Ok((tj, layer_type, json))
-    }
-
-    async fn detect_format<T>(&self, tilejson: &TileJSON, conn: &mut T) -> MbtResult<TileInfo>
+    pub async fn detect_format<T>(&self, tilejson: &TileJSON, conn: &mut T) -> MbtResult<TileInfo>
     where
         for<'e> &'e mut T: SqliteExecutor<'e>,
     {
@@ -443,234 +227,6 @@ impl Mbtiles {
         let mut conn = self.open_readonly().await?;
         self.detect_type(&mut conn).await
     }
-
-    pub async fn detect_type<T>(&self, conn: &mut T) -> MbtResult<MbtType>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        debug!("Detecting MBTiles type for {self}");
-        let typ = if is_normalized_tables_type(&mut *conn).await? {
-            MbtType::Normalized {
-                hash_view: has_tiles_with_hash(&mut *conn).await?,
-            }
-        } else if is_flat_with_hash_tables_type(&mut *conn).await? {
-            MbtType::FlatWithHash
-        } else if is_flat_tables_type(&mut *conn).await? {
-            MbtType::Flat
-        } else {
-            return Err(MbtError::InvalidDataFormat(self.filepath.clone()));
-        };
-
-        self.check_for_uniqueness_constraint(&mut *conn, typ)
-            .await?;
-
-        Ok(typ)
-    }
-
-    async fn check_for_uniqueness_constraint<T>(
-        &self,
-        conn: &mut T,
-        mbt_type: MbtType,
-    ) -> MbtResult<()>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let table_name = match mbt_type {
-            MbtType::Flat => "tiles",
-            MbtType::FlatWithHash => "tiles_with_hash",
-            MbtType::Normalized { .. } => "map",
-        };
-
-        let indexes = query("SELECT name FROM pragma_index_list(?) WHERE [unique] = 1")
-            .bind(table_name)
-            .fetch_all(&mut *conn)
-            .await?;
-
-        // Ensure there is some index on tiles that has a unique constraint on (zoom_level, tile_row, tile_column)
-        for index in indexes {
-            let mut unique_idx_cols = HashSet::new();
-            let rows = query("SELECT DISTINCT name FROM pragma_index_info(?)")
-                .bind(index.get::<String, _>("name"))
-                .fetch_all(&mut *conn)
-                .await?;
-
-            for row in rows {
-                unique_idx_cols.insert(row.get("name"));
-            }
-
-            if unique_idx_cols
-                .symmetric_difference(&HashSet::from([
-                    "zoom_level".to_string(),
-                    "tile_column".to_string(),
-                    "tile_row".to_string(),
-                ]))
-                .collect::<Vec<_>>()
-                .is_empty()
-            {
-                return Ok(());
-            }
-        }
-
-        Err(MbtError::NoUniquenessConstraint(self.filepath.clone()))
-    }
-
-    /// Perform `SQLite` internal integrity check
-    pub async fn check_integrity<T>(
-        &self,
-        conn: &mut T,
-        integrity_check: IntegrityCheckType,
-    ) -> MbtResult<()>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        if integrity_check == IntegrityCheckType::Off {
-            info!("Skipping integrity check for {self}");
-            return Ok(());
-        }
-
-        let sql = if integrity_check == IntegrityCheckType::Full {
-            "PRAGMA integrity_check;"
-        } else {
-            "PRAGMA quick_check;"
-        };
-
-        let result: Vec<String> = query(sql)
-            .map(|row: SqliteRow| row.get(0))
-            .fetch_all(&mut *conn)
-            .await?;
-
-        if result.len() > 1
-            || result.get(0).ok_or(FailedIntegrityCheck(
-                self.filepath.to_string(),
-                vec!["SQLite could not perform integrity check".to_string()],
-            ))? != "ok"
-        {
-            return Err(FailedIntegrityCheck(self.filepath().to_string(), result));
-        }
-
-        info!("{integrity_check:?} integrity check passed for {self}");
-        Ok(())
-    }
-
-    pub async fn check_agg_tiles_hashes<T>(&self, conn: &mut T) -> MbtResult<String>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let Some(stored) = self.get_agg_tiles_hash(&mut *conn).await? else {
-            return Err(AggHashValueNotFound(self.filepath().to_string()));
-        };
-        let computed = calc_agg_tiles_hash(&mut *conn).await?;
-        if stored != computed {
-            let file = self.filepath().to_string();
-            return Err(AggHashMismatch(computed, stored, file));
-        }
-
-        info!("The agg_tiles_hashes={computed} has been verified for {self}");
-        Ok(computed)
-    }
-
-    /// Compute new aggregate tiles hash and save it to the metadata table (if needed)
-    pub async fn update_agg_tiles_hash<T>(&self, conn: &mut T) -> MbtResult<String>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        let old_hash = self.get_agg_tiles_hash(&mut *conn).await?;
-        let hash = calc_agg_tiles_hash(&mut *conn).await?;
-        if old_hash.as_ref() == Some(&hash) {
-            info!("Metadata value agg_tiles_hash is already set to the correct hash `{hash}` in {self}");
-        } else {
-            if let Some(old_hash) = old_hash {
-                info!("Updating agg_tiles_hash from {old_hash} to {hash} in {self}");
-            } else {
-                info!("Adding a new metadata value agg_tiles_hash = {hash} in {self}");
-            }
-            self.set_metadata_value(&mut *conn, AGG_TILES_HASH, Some(&hash))
-                .await?;
-        }
-        Ok(hash)
-    }
-
-    pub async fn check_each_tile_hash<T>(&self, conn: &mut T) -> MbtResult<()>
-    where
-        for<'e> &'e mut T: SqliteExecutor<'e>,
-    {
-        // Note that hex() always returns upper-case HEX values
-        let sql = match self.detect_type(&mut *conn).await? {
-            MbtType::Flat => {
-                info!("Skipping per-tile hash validation because this is a flat MBTiles file");
-                return Ok(());
-            }
-            MbtType::FlatWithHash => {
-                "SELECT expected, computed FROM (
-                    SELECT
-                        upper(tile_hash) AS expected,
-                        md5_hex(tile_data) AS computed
-                    FROM tiles_with_hash
-                ) AS t
-                WHERE expected != computed
-                LIMIT 1;"
-            }
-            MbtType::Normalized { .. } => {
-                "SELECT expected, computed FROM (
-                    SELECT
-                        upper(tile_id) AS expected,
-                        md5_hex(tile_data) AS computed
-                    FROM images
-                ) AS t
-                WHERE expected != computed
-                LIMIT 1;"
-            }
-        };
-
-        query(sql)
-            .fetch_optional(&mut *conn)
-            .await?
-            .map_or(Ok(()), |v| {
-                Err(IncorrectTileHash(
-                    self.filepath().to_string(),
-                    v.get(0),
-                    v.get(1),
-                ))
-            })?;
-
-        info!("All tile hashes are valid for {self}");
-        Ok(())
-    }
-}
-
-/// Compute the hash of the combined tiles in the mbtiles file tiles table/view.
-/// This should work on all mbtiles files perf `MBTiles` specification.
-pub async fn calc_agg_tiles_hash<T>(conn: &mut T) -> MbtResult<String>
-where
-    for<'e> &'e mut T: SqliteExecutor<'e>,
-{
-    debug!("Calculating agg_tiles_hash");
-    let query = query(
-        // The md5_concat func will return NULL if there are no rows in the tiles table.
-        // For our use case, we will treat it as an empty string, and hash that.
-        // `tile_data` values must be stored as a blob per MBTiles spec
-        // `md5` functions will fail if the value is not text/blob/null
-        //
-        // Note that ORDER BY controls the output ordering, which is important for the hash value,
-        // and having it at the top level would not order values properly.
-        // See https://sqlite.org/forum/forumpost/228bb96e12a746ce
-        "
-SELECT coalesce(
-    (SELECT md5_concat_hex(
-               cast(zoom_level AS text),
-               cast(tile_column AS text),
-               cast(tile_row AS text),
-               tile_data
-           )
-           OVER (ORDER BY zoom_level, tile_column, tile_row ROWS
-                 BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING)
-     FROM tiles
-     LIMIT 1),
-    md5_hex('')
-);
-",
-    );
-    Ok(query.fetch_one(conn).await?.get::<String, _>(0))
 }
 
 pub async fn attach_hash_fn(conn: &mut SqliteConnection) -> MbtResult<()> {
@@ -684,164 +240,11 @@ pub async fn attach_hash_fn(conn: &mut SqliteConnection) -> MbtResult<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-
-    use insta::assert_yaml_snapshot;
-    use martin_tile_utils::Encoding;
-    use sqlx::Executor as _;
-    use tilejson::VectorLayer;
-
+pub(crate) mod tests {
     use super::*;
-    use crate::create_flat_tables;
 
-    async fn open(filepath: &str) -> MbtResult<(SqliteConnection, Mbtiles)> {
+    pub async fn open(filepath: &str) -> MbtResult<(SqliteConnection, Mbtiles)> {
         let mbt = Mbtiles::new(filepath)?;
         mbt.open().await.map(|conn| (conn, mbt))
-    }
-
-    #[actix_rt::test]
-    async fn mbtiles_meta() -> MbtResult<()> {
-        let filepath = "../tests/fixtures/mbtiles/geography-class-jpg.mbtiles";
-        let mbt = Mbtiles::new(filepath)?;
-        assert_eq!(mbt.filepath(), filepath);
-        assert_eq!(mbt.filename(), "geography-class-jpg");
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn metadata_jpeg() -> MbtResult<()> {
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/geography-class-jpg.mbtiles").await?;
-        let metadata = mbt.get_metadata(&mut conn).await?;
-        let tj = metadata.tilejson;
-
-        assert_eq!(tj.description.unwrap(), "One of the example maps that comes with TileMill - a bright & colorful world map that blends retro and high-tech with its folded paper texture and interactive flag tooltips. ");
-        assert!(tj.legend.unwrap().starts_with("<div style="));
-        assert_eq!(tj.maxzoom.unwrap(), 1);
-        assert_eq!(tj.minzoom.unwrap(), 0);
-        assert_eq!(tj.name.unwrap(), "Geography Class");
-        assert_eq!(tj.template.unwrap(),"{{#__location__}}{{/__location__}}{{#__teaser__}}<div style=\"text-align:center;\">\n\n<img src=\"data:image/png;base64,{{flag_png}}\" style=\"-moz-box-shadow:0px 1px 3px #222;-webkit-box-shadow:0px 1px 5px #222;box-shadow:0px 1px 3px #222;\"><br>\n<strong>{{admin}}</strong>\n\n</div>{{/__teaser__}}{{#__full__}}{{/__full__}}");
-        assert_eq!(tj.version.unwrap(), "1.0.0");
-        assert_eq!(metadata.id, "geography-class-jpg");
-        assert_eq!(metadata.tile_info, Format::Jpeg.into());
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn metadata_mvt() -> MbtResult<()> {
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/world_cities.mbtiles").await?;
-        let metadata = mbt.get_metadata(&mut conn).await?;
-        let tj = metadata.tilejson;
-
-        assert_eq!(tj.maxzoom.unwrap(), 6);
-        assert_eq!(tj.minzoom.unwrap(), 0);
-        assert_eq!(tj.name.unwrap(), "Major cities from Natural Earth data");
-        assert_eq!(tj.version.unwrap(), "2");
-        assert_eq!(
-            tj.vector_layers,
-            Some(vec![VectorLayer {
-                id: "cities".to_string(),
-                fields: vec![("name".to_string(), "String".to_string())]
-                    .into_iter()
-                    .collect(),
-                description: Some(String::new()),
-                minzoom: Some(0),
-                maxzoom: Some(6),
-                other: HashMap::default()
-            }])
-        );
-        assert_eq!(metadata.id, "world_cities");
-        assert_eq!(
-            metadata.tile_info,
-            TileInfo::new(Format::Mvt, Encoding::Gzip)
-        );
-        assert_eq!(metadata.layer_type, Some("overlay".to_string()));
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn metadata_get_key() -> MbtResult<()> {
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/world_cities.mbtiles").await?;
-
-        let res = mbt.get_metadata_value(&mut conn, "bounds").await?.unwrap();
-        assert_eq!(res, "-123.123590,-37.818085,174.763027,59.352706");
-        let res = mbt.get_metadata_value(&mut conn, "name").await?.unwrap();
-        assert_eq!(res, "Major cities from Natural Earth data");
-        let res = mbt.get_metadata_value(&mut conn, "maxzoom").await?.unwrap();
-        assert_eq!(res, "6");
-        let res = mbt.get_metadata_value(&mut conn, "nonexistent_key").await?;
-        assert_eq!(res, None);
-        let res = mbt.get_metadata_value(&mut conn, "").await?;
-        assert_eq!(res, None);
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn metadata_set_key() -> MbtResult<()> {
-        let (mut conn, mbt) = open("file:metadata_set_key_mem_db?mode=memory&cache=shared").await?;
-
-        conn.execute("CREATE TABLE metadata (name text NOT NULL PRIMARY KEY, value text);")
-            .await?;
-
-        mbt.set_metadata_value(&mut conn, "bounds", Some("0.0, 0.0, 0.0, 0.0"))
-            .await?;
-        assert_eq!(
-            mbt.get_metadata_value(&mut conn, "bounds").await?.unwrap(),
-            "0.0, 0.0, 0.0, 0.0"
-        );
-
-        mbt.set_metadata_value(
-            &mut conn,
-            "bounds",
-            Some("-123.123590,-37.818085,174.763027,59.352706"),
-        )
-        .await?;
-        assert_eq!(
-            mbt.get_metadata_value(&mut conn, "bounds").await?.unwrap(),
-            "-123.123590,-37.818085,174.763027,59.352706"
-        );
-
-        mbt.set_metadata_value(&mut conn, "bounds", None).await?;
-        assert_eq!(mbt.get_metadata_value(&mut conn, "bounds").await?, None);
-
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn detect_type() -> MbtResult<()> {
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/world_cities.mbtiles").await?;
-        let res = mbt.detect_type(&mut conn).await?;
-        assert_eq!(res, MbtType::Flat);
-
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/zoomed_world_cities.mbtiles").await?;
-        let res = mbt.detect_type(&mut conn).await?;
-        assert_eq!(res, MbtType::FlatWithHash);
-
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/geography-class-jpg.mbtiles").await?;
-        let res = mbt.detect_type(&mut conn).await?;
-        assert_eq!(res, MbtType::Normalized { hash_view: false });
-
-        let (mut conn, mbt) = open(":memory:").await?;
-        let res = mbt.detect_type(&mut conn).await;
-        assert!(matches!(res, Err(MbtError::InvalidDataFormat(_))));
-
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn validate_valid_file() -> MbtResult<()> {
-        let (mut conn, mbt) = open("../tests/fixtures/mbtiles/zoomed_world_cities.mbtiles").await?;
-        mbt.check_integrity(&mut conn, IntegrityCheckType::Quick)
-            .await?;
-        Ok(())
-    }
-
-    #[actix_rt::test]
-    async fn validate_invalid_file() -> MbtResult<()> {
-        let (mut conn, mbt) =
-            open("../tests/fixtures/files/invalid_zoomed_world_cities.mbtiles").await?;
-        let result = mbt.check_agg_tiles_hashes(&mut conn).await;
-        assert!(matches!(result, Err(MbtError::AggHashMismatch(..))));
-        Ok(())
     }
 }
